Refactor list commands and list parsers
