+++ src/main/java/seedu/ddd/logic/commands/list/ListVendorCommand.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.commands;
+package seedu.ddd.logic.commands.list;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.ddd.logic.Messages.MESSAGE_CONTACTS_LISTED_OVERVIEW;
@@ -6,15 +6,16 @@ import static seedu.ddd.logic.Messages.MESSAGE_CONTACTS_LISTED_OVERVIEW;
 import java.util.function.Predicate;
 
 import seedu.ddd.commons.util.ToStringBuilder;
+import seedu.ddd.logic.commands.CommandResult;
 import seedu.ddd.model.Model;
 import seedu.ddd.model.contact.common.Contact;
 
 /**
- * Lists clients in the address book to the user.
+ * Lists vendors in the address book to the user.
  */
-public class ListClientCommand extends ListContactCommand {
+public class ListVendorCommand extends ListContactCommand {
 
-    public ListClientCommand(Predicate<Contact> predicate) {
+    public ListVendorCommand(Predicate<Contact> predicate) {
         super(predicate);
     }
 
@@ -32,11 +33,11 @@ public class ListClientCommand extends ListContactCommand {
             return true;
         }
         // instanceof handles nulls
-        if (!(other instanceof ListClientCommand)) {
+        if (!(other instanceof ListVendorCommand)) {
             return false;
         }
-        ListClientCommand otherListClientCommand = (ListClientCommand) other;
-        return this.getPredicate().equals(otherListClientCommand.getPredicate());
+        ListVendorCommand otherListVendorCommand = (ListVendorCommand) other;
+        return this.getPredicate().equals(otherListVendorCommand.getPredicate());
     }
 
     @Override
