+++ src/main/java/seedu/ddd/logic/parser/list/ListContactCommandParser.java
@@ -1,7 +1,6 @@
-package seedu.ddd.logic.parser;
+package seedu.ddd.logic.parser.list;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.ddd.logic.parser.CliFlags.FLAG_VENDOR;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_ADDRESS;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_DESC;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_EMAIL;
@@ -10,26 +9,28 @@ import static seedu.ddd.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_TAG;
 
-import seedu.ddd.logic.commands.ListVendorCommand;
+import seedu.ddd.logic.commands.list.ListContactCommand;
+import seedu.ddd.logic.parser.ArgumentMultimap;
+import seedu.ddd.logic.parser.ArgumentTokenizer;
+import seedu.ddd.logic.parser.Parser;
 import seedu.ddd.logic.parser.exceptions.ParseException;
 import seedu.ddd.model.contact.common.predicate.ContactPredicateBuilder;
-import seedu.ddd.model.contact.common.predicate.VendorPredicateBuilder;
 
 /**
- * Parses input arguments and creates a new ListVendorCommand object
+ * Parses input arguments and creates a new ListContactCommand object
  */
-public class ListVendorCommandParser implements Parser<ListVendorCommand> {
+public class ListContactCommandParser implements Parser<ListContactCommand> {
     /**
-     * Parses the given {@code String} of arguments in the context of the ListVendorCommand
-     * and returns a ListVendorCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the ListContactCommand
+     * and returns a ListCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public ListVendorCommand parse(String args) throws ParseException {
+    public ListContactCommand parse(String args) throws ParseException {
         requireNonNull(args);
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_ID, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL,
-                        PREFIX_ADDRESS, PREFIX_TAG, PREFIX_DESC, FLAG_VENDOR);
-        ContactPredicateBuilder predicateBuilder = new VendorPredicateBuilder(argMultimap);
-        return new ListVendorCommand(predicateBuilder.build());
+                        PREFIX_ADDRESS, PREFIX_TAG, PREFIX_DESC);
+        ContactPredicateBuilder combinedPredicate = new ContactPredicateBuilder(argMultimap);
+        return new ListContactCommand(combinedPredicate.build());
     }
 }
