+++ src/main/java/seedu/ddd/logic/parser/list/ListEventCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.ddd.logic.parser;
+package seedu.ddd.logic.parser.list;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_ADDRESS;
@@ -9,25 +9,28 @@ import static seedu.ddd.logic.parser.CliSyntax.PREFIX_NAME;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_PHONE;
 import static seedu.ddd.logic.parser.CliSyntax.PREFIX_TAG;
 
-import seedu.ddd.logic.commands.ListContactCommand;
+import seedu.ddd.logic.commands.list.ListEventCommand;
+import seedu.ddd.logic.parser.ArgumentMultimap;
+import seedu.ddd.logic.parser.ArgumentTokenizer;
+import seedu.ddd.logic.parser.Parser;
 import seedu.ddd.logic.parser.exceptions.ParseException;
-import seedu.ddd.model.contact.common.predicate.ContactPredicateBuilder;
+import seedu.ddd.model.event.common.predicate.EventPredicateBuilder;
 
 /**
- * Parses input arguments and creates a new ListContactCommand object
+ * Parses input arguments and creates a new ListEventCommand object
  */
-public class ListContactCommandParser implements Parser<ListContactCommand> {
+public class ListEventCommandParser implements Parser<ListEventCommand> {
     /**
-     * Parses the given {@code String} of arguments in the context of the ListContactCommand
-     * and returns a ListCommand object for execution.
+     * Parses the given {@code String} of arguments in the context of the ListEventCommand
+     * and returns a ListEventCommand object for execution.
      * @throws ParseException if the user input does not conform the expected format
      */
-    public ListContactCommand parse(String args) throws ParseException {
+    public ListEventCommand parse(String args) throws ParseException {
         requireNonNull(args);
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_ID, PREFIX_NAME, PREFIX_PHONE, PREFIX_EMAIL,
                         PREFIX_ADDRESS, PREFIX_TAG, PREFIX_DESC);
-        ContactPredicateBuilder combinedPredicate = new ContactPredicateBuilder(argMultimap);
-        return new ListContactCommand(combinedPredicate.build());
+        EventPredicateBuilder predicateBuilder = new EventPredicateBuilder(argMultimap);
+        return new ListEventCommand(predicateBuilder.build());
     }
 }
