+++ src/main/java/seedu/ddd/model/contact/common/predicate/ContactPredicateBuilder.java
@@ -13,7 +13,7 @@ import java.util.Set;
 import java.util.function.Predicate;
 
 import seedu.ddd.commons.util.ToStringBuilder;
-import seedu.ddd.logic.commands.ListCommand;
+import seedu.ddd.logic.commands.list.ListCommand;
 import seedu.ddd.logic.parser.ArgumentMultimap;
 import seedu.ddd.logic.parser.ParserUtil;
 import seedu.ddd.logic.parser.exceptions.ParseException;
