+++ src/main/java/seedu/ddd/model/event/common/predicate/EventPredicateBuilder.java
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import java.util.function.Predicate;
 
 import seedu.ddd.commons.util.ToStringBuilder;
-import seedu.ddd.logic.commands.ListEventCommand;
+import seedu.ddd.logic.commands.list.ListEventCommand;
 import seedu.ddd.logic.parser.ArgumentMultimap;
 import seedu.ddd.logic.parser.exceptions.ParseException;
 import seedu.ddd.model.common.Id;
