+++ src/test/java/seedu/ddd/logic/commands/ListEventCommandTest.java
@@ -15,6 +15,7 @@ import java.util.Collections;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import seedu.ddd.logic.commands.list.ListEventCommand;
 import seedu.ddd.model.Model;
 import seedu.ddd.model.ModelManager;
 import seedu.ddd.model.UserPrefs;
