+++ src/test/java/seedu/ddd/logic/parser/AddressBookParserTest.java
@@ -16,9 +16,9 @@ import seedu.ddd.logic.commands.EditContactCommand;
 import seedu.ddd.logic.commands.EditContactCommand.EditContactDescriptor;
 import seedu.ddd.logic.commands.ExitCommand;
 import seedu.ddd.logic.commands.HelpCommand;
-import seedu.ddd.logic.commands.ListCommand;
-import seedu.ddd.logic.commands.ListContactCommand;
-import seedu.ddd.logic.commands.ListEventCommand;
+import seedu.ddd.logic.commands.list.ListCommand;
+import seedu.ddd.logic.commands.list.ListContactCommand;
+import seedu.ddd.logic.commands.list.ListEventCommand;
 import seedu.ddd.logic.parser.exceptions.ParseException;
 import seedu.ddd.model.contact.common.Contact;
 import seedu.ddd.testutil.EditContactDescriptorBuilder;
